fixed the rest of responsiveness

+++ src/Components/Card.jsx
@@ -3,10 +3,10 @@ import { MdOutlineDeleteOutline } from "react-icons/md";
 import { FaUserEdit } from "react-icons/fa";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
-// import { IoMenu } from "react-icons/io5";
 
 function Card() {
   const [data, setData] = useState([]);
+  const [error, setError] = useState(null);
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -17,17 +17,23 @@ function Card() {
       })
       .catch((error) => {
         console.error("Error fetching data:", error);
+        setError(error);
       });
   }, []);
 
   const handleDeleteClick = async (id) => {
-    try {
-      await axios.delete(
-        `${process.env.REACT_APP_API_BASE_URL}/employee/${id}`
-      );
-      setData((prevData) => prevData.filter((employee) => employee.id !== id));
-    } catch (error) {
-      console.error("Error during deletion:", error);
+    if (window.confirm("Are you sure you want to delete this employee?")) {
+      try {
+        await axios.delete(
+          `${process.env.REACT_APP_API_BASE_URL}/employee/${id}`
+        );
+        setData((prevData) =>
+          prevData.filter((employee) => employee.id !== id)
+        );
+      } catch (error) {
+        console.error("Error during deletion:", error);
+        setError(error);
+      }
     }
   };
 
@@ -35,13 +41,17 @@ function Card() {
     navigate(`/employee/edit/${id}`);
   };
 
+  if (error) {
+    return <div>Error: {error.message}</div>;
+  }
+
   return (
-    <div>
-      <div className="flex flex-wrap justify-center ">
+    <div className="bg-slate-50">
+      <div className="flex flex-wrap justify-center w-full">
         {data.map((employee) => (
           <div
             key={employee.id}
-            className="m-4 p-6 bg-white rounded-md shadow-md w-[15%] max-sm:w-[80%]  max-sm:block max-md:w-[40%] max-lg:w-[40%]  "
+            className="m-4 p-6 bg-white rounded-md shadow-md w-[15%] max-sm:w-[80%]  max-sm:block  max-md:w-[40%]  max-lg:w-[40%] ... hover:shadow-xl hover:scale-105 transition-transform duration-300 ease-in-out "
           >
             <div className="mb-4 flex justify-center">
               <img
@@ -51,26 +61,28 @@ function Card() {
               />
             </div>
             <div className="w-auto">
-              <p className="text-lg font-semibold">
+              <p className="text-lg font-semibold max-md:text-base md:text-base">
                 {`${employee.first_name} ${employee.last_name}`}
               </p>
-              <p className="text-gray-600 text-xs flex-grow">
+
+              <p className="text-gray-600 text-xs underline max-md:text-xs md:text-[0.7rem] ">
                 {employee.email}
               </p>
+
               <p className="text-gray-600">{employee.number}</p>
               <p className="text-gray-600">
                 {employee.gender === "F" ? "Female" : "Male"}
               </p>
             </div>
-            <div className="mt-[-23px] space-x-2 justify-end flex">
+            <div className="mt-[-23px] space-x-2 justify-end flex ">
               <button
-                className="bg-red-500 text-white p-2 rounded-full"
+                className="bg-red-500 text-white p-2 rounded-full ... hover:scale-105 transition-transform duration-300 ease-in-out"
                 onClick={() => handleDeleteClick(employee.id)}
               >
                 <MdOutlineDeleteOutline />
               </button>
               <button
-                className="bg-green-500 text-white  p-2 rounded-full"
+                className="bg-green-500 text-white  p-2 rounded-full ... hover:scale-105 transition-transform duration-300 ease-in-out"
                 onClick={() => handelEditClick(employee.id)}
               >
                 <FaUserEdit />
